docs/requirements.txt
PyYAML==6.0.3
mkdocs==1.6.1
mkdocs-material==9.6.21
mkdocs-macros-plugin==1.4.0